pyproject.toml
[project]
name = "cloud"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "KmasaK", email = "155167903+KmasaK@users.noreply.github.com" }
]
dependencies = [
    "joblib==1.3.2",
    "numpy==1.23.4",
    "pandas==2.2.1",
    "python-dateutil==2.8.2",
    "pytz==2024.1",
    "scikit-learn==1.5.1",
    "scipy==1.12.0",
    "tzdata==2024.1",
    "six==1.16.0",
    "sklearn-pandas==2.2.0",
    "threadpoolctl==3.3.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
source = ["cloud"]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["cloud"]
